perform handshake from a content script rather than with scripting
In this commit, we remove the need for the background script to inject a
function, and instead place a content script on every page the extension
is active that the background script connects to in order to begin the flow.

Here is a diagram of the flow now:

1. Background              2. Content Script     3. Content Script       4. SSO
                              (current tab)         (callback url)

tabs.connect(...)   ------>   Create iFrame  -------------┐
                                                          |
runtime.onConnect(...)  <---------------------    runtime.connect(...)
     |                   Asks for /authorize url
     |
postMessage(...)  ----------------------------->  onMessage(...) ------> iframe
                                                                           |
                                                                           |
disconnect() <---------------------------------   postMessage(...) --------┘
                      Code & State

We also remove the need for the end user to add any code to (3),
since by setting the new content script to be injected in all frames, it
will also be injectend into the new iFrame that is created by (2). This
also nicely matches with where the user of this library will need to get
authentication tokens, as if they are performing an operation from a
content script, then it will mean we will be able to grab a token from
the same page that the user is currently active on which we are
guaranteed to have access to.