add support for in memory caching of access tokens
This commit adds the ability to save retrieved access tokens into
memory, and when subsequent calls are made to getTokenSilently, these
tokens will be returned if they are still valid. These will only last
for the life of the service worker, and if the service worker shuts down
and restart, the full _getTokenWithIFrame will need to be performed. I
decided not to add support for localstorage cache location as the security of
chrome.storage.local is dubious, but this may be something
that can be added in the future.